requirements.txt
crewai>=0.86.0
crewai-tools>=0.17.0
pydantic>=2.6.0
python-dotenv>=1.0.1
wikipedia>=1.4.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
streamlit>=1.33.0
kagglehub>=0.3.0
ipykernel>=6.29.0
jupyter>=1.0.0
tqdm>=4.66.0
